rework storage etc
Smarter approach to committing to the github branch. The files to commit
are built up in a tmp dir, and only at the end does it switch to the
branch, stage and commit all the files. Eventually it might be nice to
avoid switching branches at all when committing. I've done that before in
eg, pristine-tar.

So, it no longer needs to store every request and every filename. The
memory use had been alarmingly high, this should improve it. Only failed
requests are stored.

The tricky thing involved making sure requests don't get re-run. Turns
out that the only things that could re-run requests are:

* retried requests that later come up again. So keep a set of the retried
  requests that got run, and avoid re-running those.
* fork discovery, in the unlikely case that there are cycles in github's
  fork data. Guarded against by not recursing if the fork is already setup
  as a git remote.