fix memory leak
Oddly, re-reading the git config caused a memory leak. Profiling showed
all memory was being used by list conses. Presumably for the list of
remotes in the git config. Although it grew too fast to just be that;
at an increasing rate. My guess is that, somehow, the old git config
data was retained. However, I cannot see anything that would cause that
to happen, since the value is stored in the monad. Nothing should refer
to an old value of that state. Weird.

Anyway, to fix it, let's just not reread git config. This does mean that
if github puts in a fork loop, we might go round and round. It also
complicated fetching from the remotes. Now we fetch when adding a new fork,
and all the existing forks are fetched from at startup.